% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{act_file_download}
\alias{act_file_download}
\title{Place where functions live for the matos_*_summary family of functions}
\usage{
act_file_download(
  type,
  temp_dir = NULL,
  matos_project = NULL,
  project_files = NULL
)
}
\arguments{
\item{type}{type of data: qualified or unqualified detections; or deployment metadata}

\item{temp_dir}{location of temporary directory}

\item{matos_project}{matos project number}

\item{project_files}{qualified/unqualified detection files}
}
\description{
Place where functions live for the matos_*_summary family of functions
}
\keyword{internal}
