% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_projects.R
\name{flatten_names}
\alias{flatten_names}
\title{Flatten and remove special characters to aid in matching}
\usage{
flatten_names(x)
}
\arguments{
\item{x}{name to flatten}
}
\description{
Flatten and remove special characters to aid in matching
}
\keyword{internal}
