% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_projects.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Use \code{agrep} for fuzzy matching}
\usage{
fuzzy_match(a, b)
}
\arguments{
\item{a}{vector holding names that may be a part of \code{b} by changing less than
25\% of the characters}

\item{b}{vector holding names that may encompass some manipulation of \code{a}}
}
\description{
Use \code{agrep} for fuzzy matching
}
\keyword{internal}
