% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_otn_template.R
\name{get_otn_template}
\alias{get_otn_template}
\title{Download Ocean-Tracking-Network-style metadata templates}
\usage{
get_otn_template(
  template_type = c("receiver", "tag", "glider"),
  out_dir = NULL,
  quiet = FALSE
)
}
\arguments{
\item{template_type}{Character string. One of: "receiver" (default), the
deployment data submittal template; "tag", the tagging data submittal
template; or "glider", the wave and Slocum glider metadata template.}

\item{out_dir}{Optional character string noting where you would like the file
to be downloaded. Defaults to the working directory.}

\item{quiet}{suppress status messages from \code{download.file}}
}
\value{
Ocean Tracking Network metadata template in XLSX format.
}
\description{
Download Ocean-Tracking-Network-style metadata templates
}
\examples{
\dontrun{
# Tag metadata template downloaded to working directory
get_otn_template()

# Glider metadata template downloaded to temporary directory
get_otn_template("glider", out_dir = tempdir())
}
}
