% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_my_projects.R
\name{list_my_projects_mem}
\alias{list_my_projects_mem}
\title{Memoised list_my_projects function for internal use}
\usage{
list_my_projects_mem(read_access, warn_multimatch)
}
\arguments{
\item{read_access}{Do you want to only list projects for which you have
file-read permission? Defaults to TRUE, though there is significant
speed up if switched to FALSE.}

\item{warn_multimatch}{Warn you if there have been multiple project matches?
Defaults to TRUE.}
}
\description{
Memoised list_my_projects function for internal use
}
\keyword{internal}
