% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_projects.R
\name{list_projects}
\alias{list_projects}
\title{List MATOS projects}
\usage{
list_projects(
  what = c("all", "mine"),
  read_access = TRUE,
  quiet = FALSE,
  force = FALSE
)
}
\arguments{
\item{what}{What list of projects do you want returned: all projects ("all",
default) or your projects ("mine")?}

\item{read_access}{If listing your projects, do you want to only list projects
for which you have file-read permission? Defaults to TRUE, though there
is significant speed up if switched to FALSE.}

\item{quiet}{Do you want to suppress messages regariding matched projects?}

\item{force}{Do you want to reset the cache and re-ping the database?}
}
\description{
This function scrapes the table found at \url{https://matos.asascience.com/project}
and combines it with project metadata stored on the
\href{https://members.oceantrack.org/geoserver/web/}{Ocean Tracking Network Geoserver}.
This table provides the full name of the project, collection code, MATOS
project number, MATOS project page URL, project status, full name, citation,
website, project type, area, and abstract. You do not need to log in
via \code{matos_login} or have any permissions to view/download this table.
}
\examples{
\dontrun{
# List all projects, the default:
list_projects()

# List your projects (which may contain some for which you do not have read access):
list_projects("mine", read_access = F)
}
}
