% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_projects.R
\name{list_projects_mem}
\alias{list_projects_mem}
\title{Memoised list_projects function for internal use.}
\usage{
list_projects_mem(what, read_access, quiet)
}
\arguments{
\item{what}{What list of projects do you want returned: all projects ("all",
default) or your projects ("mine")?}

\item{read_access}{If listing your projects, do you want to only list
projects for which you have file-read permission? Defaults to TRUE,
though there is significant speed up if switched to FALSE.}

\item{quiet}{Do you want to suppress messages regarding matched projects?
Defaults to FALSE.}
}
\description{
Memoised list_projects function for internal use.
}
\keyword{internal}
