% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matos_tag_summary.R
\name{matos_tag_summary}
\alias{matos_tag_summary}
\title{Create summary reports of receiver project data from the OTN data push}
\usage{
matos_tag_summary(matos_project = NULL, matched = NULL, ...)
}
\arguments{
\item{matos_project}{MATOS project number or name that you wish to have summarized}

\item{matched}{Default is NULL: OTN matched detections will be downloaded from MATOS and unzipped. If you do not wish to download your files, this argument also accepts a character vector of file paths of your matched detections. These can be CSVs or zipped folders.}

\item{...}{Arguments passed to \code{otndo::make_tag_push_summary}}
}
\description{
Create summary reports of receiver project data from the OTN data push
}
\examples{
\dontrun{
# You can just use your ACT project number
matos_tag_summary(87)

# Or provide an optional date to summarize "What's New".
matos_tag_summary(87, since = "2018-11-01")
}
}
