% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_projects.R
\name{within_match}
\alias{within_match}
\title{Check for projects within other project names}
\usage{
within_match(a, b)
}
\arguments{
\item{a}{vector holding names that may be a part of \code{b}}

\item{b}{vector holding names that may encompass \code{a}}
}
\description{
Check for projects within other project names
}
\keyword{internal}
